[Xposed] Don't crash for dex files without extension
DexFilenameToOdexFilename assumes that all dex files have a three-letter
extension, but that is not always true. In particular, the app "Quizduell"
creates temporary dex files without extension (and non-alphabetic characters)
and crashes during startup.

Xposed uses this function in more places than ART itself so this assumption
becomes visible. DexFile.isDexOptNeeded() would probably crash as well if it
was called.